+++ airbyte-cdk/bulk/core/extract/src/main/kotlin/io/airbyte/cdk/output/sockets/SocketOutputConsumer.kt
@@ -10,6 +10,7 @@ import io.airbyte.cdk.util.Jsons
 import io.airbyte.protocol.models.v0.AirbyteMessage
 import io.airbyte.protocol.models.v0.AirbyteRecordMessage
 import io.airbyte.protocol.models.v0.AirbyteRecordMessageMeta
+import io.airbyte.protocol.protobuf.AirbyteMessage.AirbyteMessageProtobuf
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.context.annotation.Factory
 import io.micronaut.context.annotation.Requires
@@ -26,13 +27,13 @@ class BoostedOutputConsumerFactory(
     val bufferByteSizeThresholdForFlush: Int,
 ) {
 
-    fun boostedOutputConsumer(socket: SocketWrapper, additionalProperties: Map<String, String>): BoostedOutputConsumer {
-        return BoostedOutputConsumer(socket, clock, bufferByteSizeThresholdForFlush, additionalProperties) // TEMP
+    fun boostedOutputConsumer(socket: SocketWrapper, additionalProperties: Map<String, String>): SocketJsonOutputConsumer {
+        return SocketJsonOutputConsumer(socket, clock, bufferByteSizeThresholdForFlush, additionalProperties) // TEMP
     }
 }
 
-class BoostedOutputConsumer(
-    /*private*/ val socket: SocketWrapper,
+class SocketJsonOutputConsumer(
+    private val socket: SocketWrapper,
     clock: Clock,
     val bufferByteSizeThresholdForFlush: Int,
     private val additionalProperties: Map<String, String>,
@@ -165,4 +166,76 @@ class BoostedOutputConsumer(
         const val META_PREFIX = ""","meta":"""
     }
 
+}
+
+class SocketProtobufOutputConsumer(
+    private val socket: SocketWrapper,
+    private val clock: Clock,
+    val bufferByteSizeThresholdForFlush: Int,
+): OutputConsumer(clock)
+{
+    private val log = KotlinLogging.logger {}
+    private val buffer = ByteArrayOutputStream()
+
+    override fun accept(airbyteMessage: io.airbyte.protocol.models.v0.AirbyteMessage) {
+        // This method effectively println's its JSON-serialized argument.
+        // Using println is not particularly efficient, however.
+        // To improve performance, this method accumulates RECORD messages into a buffer
+        // before writing them to standard output in a batch.
+        if (airbyteMessage.type == io.airbyte.protocol.models.v0.AirbyteMessage.Type.RECORD) {
+            // RECORD messages undergo a different serialization scheme.
+            //accept(airbyteMessage.record)
+            //no-op
+        } else {
+            synchronized(this) {
+                val b = ByteArrayOutputStream()
+                val sequenceWriter: SequenceWriter = Jsons.writer().writeValues(b)
+                sequenceWriter.write(airbyteMessage)
+                sequenceWriter.flush()
+
+                val pm = AirbyteMessageProtobuf.newBuilder()
+                    .setAirbyteProtocolMessage(String(b.toByteArray()))
+                    .build()
+                pm.writeDelimitedTo(buffer)
+
+//                // Write a newline character to the buffer if it's not empty.
+//                withLockMaybeWriteNewline()
+//                // Non-RECORD AirbyteMessage instances are serialized and written to the buffer
+//                // using standard jackson object mapping facilities.
+//                sequenceWriter.write(airbyteMessage)
+//                sequenceWriter.flush()
+//                // Such messages don't linger in the buffer, they are flushed to stdout immediately,
+//                // along with whatever might have already been lingering inside.
+//                // This prints a newline after the message.
+                log.info { "AirbyteMessage sent over SOCKET: ${String(b.toByteArray())}" }
+                withLockFlush()
+            }
+        }
+    }
+    fun accept(airbyteProtoMessage: AirbyteMessageProtobuf) {
+        synchronized(this) {
+            airbyteProtoMessage.writeDelimitedTo(buffer)
+            if (buffer.size() >= bufferByteSizeThresholdForFlush) {
+                withLockFlush()
+            }
+
+        }
+     }
+
+    private fun withLockFlush() {
+        if (buffer.size() > 0) {
+            buffer.writeTo(socket.outputStream)
+//            socket.outputStream?.write(System.lineSeparator().toByteArray())
+//            stdout.println(buffer.toString(Charsets.UTF_8))
+//            stdout.flush()
+            buffer.reset()
+        }
+    }
+
+    override fun close() {
+        synchronized(this) {
+            // Flush any remaining buffer contents to stdout before closing.
+            withLockFlush()
+        }
+    }
 }
